requirements.txt
# SPF/DMARC/DKIM Watcher – Python dependencies
# Install with: pip install -r requirements.txt

# Web framework
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.0

# WSGI / security
Werkzeug>=3.1.0
bcrypt>=4.0.0

# DNS checking
dnspython>=2.7.0
checkdmarc>=5.14.0

# Cryptography (DKIM key validation)
cryptography>=42.0.0

# IP reputation / DNSBL
pydnsbl>=1.1.7

# WHOIS domain registrar lookup
python-whois>=0.9.4

# Timezone data (required on Windows; bundled on Linux/macOS)
tzdata>=2024.1

# WTForms (pulled in by Flask-WTF; listed explicitly for clarity)
WTForms>=3.1.0

# Production WSGI server
gunicorn>=22.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
